+++ include/crossedfingers/internals/display/OutputWrapper.h
@@ -21,40 +21,30 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-#ifndef DEFAULTDISPLAY_H
-#define DEFAULTDISPLAY_H
+#ifndef OUTPUTWRAPPER_H
+#define OUTPUTWRAPPER_H
 /**
- * Default simple display
+ * Wrapper around iostream
  */
 
-#include "Display.h"
+#include <sstream>
+#include <string>
 
-namespace crossedfingers {
-class DefaultDisplay final : public Display {
+namespace crossedfingers::internals {
+class OutputWrapper final {
   public:
-    auto printBeginSuite(const std::string &suite_name) -> void override;
+    static auto init() -> OutputWrapper &;
 
-    auto printBeginCase(const std::string &case_name) -> void override;
+    static auto print(const std::string &str) noexcept -> void;
 
-    auto printEndCase(const std::string &case_name) -> void override;
+  private:
+    std::streambuf *_out;
+    std::stringstream _redirection;
 
-    auto printEndSuite(const std::string &suite_name) -> void override;
+    OutputWrapper();
 
-    auto printSkipCase(const std::string &case_name) -> void override;
-
-    auto printWarningCase(const std::string &case_name) -> void override;
-
-    auto printFailCase(const std::string &case_name, const std::string &message) -> void override;
-
-    auto printSummary(
-        int test_count,
-        int assertion_count,
-        const std::vector<std::string> &succeed_tests,
-        const std::vector<std::string> &skipped_tests,
-        const std::map<std::string, std::string> &warning_tests,
-        const std::map<std::string, std::string> &failed_tests
-    ) -> void override;
+    ~OutputWrapper();
 };
 } // namespace crossedfingers
 
-#endif // DEFAULTDISPLAY_H
+#endif // OUTPUTWRAPPER_H
