+++ src/crossedfingers/TestCase.cpp
@@ -21,13 +21,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-#include "crossedfingers/TestCase.h"
+#include "crossedfingers/internals/TestCase.h"
 
-#include "crossedfingers/display/OutputWrapper.h"
+#include "crossedfingers/internals/display/OutputWrapper.h"
 
 #include <utility>
 
-using namespace crossedfingers;
+using namespace crossedfingers::internals;
 
 TestCase::TestCase(std::string name, const std::function<void()> &callback)
     : _name(std::move(name)), _callback(callback) {}
